+++ module/sheets/item-sheet.mjs
@@ -1,3 +1,5 @@
+import { mergeSheetEditLockContext, wireSheetEditLock } from "./sheet-edit-lock.mjs";
+
 const { ItemSheetV2 } = foundry.applications.sheets;
 const { HandlebarsApplicationMixin } = foundry.applications.api;
 
@@ -18,7 +20,7 @@ export class PTGItemSheet extends HandlebarsApplicationMixin(ItemSheetV2) {
   };
 
   async _prepareContext(options) {
-    const context = await super._prepareContext(options);
+    const context = mergeSheetEditLockContext(await super._prepareContext(options), this, this.item);
 
     context.item = this.item;
     context.system = this.item.system;
@@ -26,6 +28,11 @@ export class PTGItemSheet extends HandlebarsApplicationMixin(ItemSheetV2) {
     context.config = CONFIG.PTG;
 
     return context;
+  }
+
+  async _onRender(context, options) {
+    await super._onRender(context, options);
+    wireSheetEditLock(this, this.element, this.item);
   }
 }